backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
python-dotenv
groq
pymongo